# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MusicBox API",
        "description": "Documentation interactive de l'API MusicBox (Artistes, Albums, Chansons)",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/songs": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Lister toutes les chansons",
                "description": "Retourne la liste paginée des chansons",
                "operationId": "0388bb28c35eec49b8b492d169ce3575",
                "responses": {
                    "200": {
                        "description": "Succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "title": {
                                                "type": "string",
                                                "example": "Billie Jean"
                                            },
                                            "album_id": {
                                                "type": "integer",
                                                "example": 1
                                            },
                                            "duration": {
                                                "type": "string",
                                                "example": "04:15"
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Songs"
                ],
                "summary": "Créer une nouvelle chanson",
                "description": "Ajoute une chanson dans la base de données",
                "operationId": "2f1970b82d91cc715578d972189f19cb",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "title",
                                    "album_id",
                                    "duration"
                                ],
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Billie Jean"
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "04:15"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Chanson créée avec succès",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Billie Jean"
                                        },
                                        "album_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "duration": {
                                            "type": "string",
                                            "example": "04:15"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/songs/{id}": {
            "put": {
                "tags": [
                    "Songs"
                ],
                "summary": "Mettre à jour une chanson",
                "description": "Met à jour les informations d’une chanson existante",
                "operationId": "70cdd4af6ae4deea2a031741a1679cf5",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la chanson",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Beat It"
                                    },
                                    "album_id": {
                                        "type": "integer",
                                        "example": 1
                                    },
                                    "duration": {
                                        "type": "string",
                                        "example": "04:18"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Chanson mise à jour",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "title": {
                                            "type": "string",
                                            "example": "Beat It"
                                        },
                                        "album_id": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "duration": {
                                            "type": "string",
                                            "example": "04:18"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Chanson non trouvée"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Songs"
                ],
                "summary": "Supprimer une chanson",
                "description": "Supprime une chanson de la base de données",
                "operationId": "829c1a52ae5013ce3ff4b9829a31485f",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID de la chanson",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Chanson supprimée avec succès"
                    },
                    "404": {
                        "description": "Chanson non trouvée"
                    }
                }
            }
        },
        "/api/ad/songs/search": {
            "get": {
                "tags": [
                    "Songs"
                ],
                "summary": "Rechercher des chansons",
                "description": "Recherche des chansons par titre ou par artiste (avec pagination)",
                "operationId": "cd59d50d5bbc060ff87cd770b2c97fe7",
                "parameters": [
                    {
                        "name": "query",
                        "in": "query",
                        "description": "Mot-clé à rechercher (titre de la chanson ou nom de l'artiste)",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "example": "Queen"
                        }
                    },
                    {
                        "name": "page",
                        "in": "query",
                        "description": "Numéro de page pour la pagination",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "example": 1
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Résultats trouvés",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "current_page": {
                                            "type": "integer",
                                            "example": 1
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "properties": {
                                                    "id": {
                                                        "type": "integer",
                                                        "example": 1
                                                    },
                                                    "title": {
                                                        "type": "string",
                                                        "example": "Bohemian Rhapsody"
                                                    },
                                                    "album": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 10
                                                            },
                                                            "name": {
                                                                "type": "string",
                                                                "example": "A Night at the Opera"
                                                            },
                                                            "artist": {
                                                                "properties": {
                                                                    "id": {
                                                                        "type": "integer",
                                                                        "example": 3
                                                                    },
                                                                    "name": {
                                                                        "type": "string",
                                                                        "example": "Queen"
                                                                    }
                                                                },
                                                                "type": "object"
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                },
                                                "type": "object"
                                            }
                                        },
                                        "total": {
                                            "type": "integer",
                                            "example": 50
                                        },
                                        "per_page": {
                                            "type": "integer",
                                            "example": 10
                                        },
                                        "last_page": {
                                            "type": "integer",
                                            "example": 5
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Paramètre query manquant"
                    },
                    "404": {
                        "description": "Aucun résultat trouvé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
...
}